♻️ Move auth
