+++ apps/web/src/auth/merge-user.ts
