+++ apps/web/src/pages/api/auth/[...nextauth].ts
@@ -1,13 +1,13 @@
 import type { NextApiRequest, NextApiResponse } from "next";
 
 import { posthogApiHandler } from "@/app/posthog";
-import { AuthApiRoute } from "@/utils/auth";
+import { AuthApiRoute } from "@/auth";
 import { composeApiHandlers } from "@/utils/next";
 
 export default async function auth(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === "HEAD") {
     res.status(200).end();
-    res.setHeader('Content-Length', '0');
+    res.setHeader("Content-Length", "0");
   } else {
     return composeApiHandlers(AuthApiRoute, posthogApiHandler)(req, res);
   }
