+++ apps/web/src/pages/api/trpc/[trpc].ts
@@ -4,7 +4,7 @@ import { createNextApiHandler } from "@trpc/server/adapters/next";
 
 import { posthogApiHandler } from "@/app/posthog";
 import { AppRouter, appRouter } from "@/trpc/routers";
-import { getServerSession } from "@/utils/auth";
+import { getServerSession } from "@/auth";
 import { getEmailClient } from "@/utils/emails";
 import { composeApiHandlers } from "@/utils/next";
